Initial structure to use emlib and centralize import xmippLib in a single place
