+++ pwem/emlib/__init__.py
@@ -1,12 +1,12 @@
 # **************************************************************************
 # *
-# * Authors:     David Maluenda Niubo (dmaluenda@cnb.csic.es)
+# * Authors:     J.M. de la Rosa Trevin (delarosatrevin@scilifelab.se) [1]
 # *
-# * Unidad de  Bioinformatica of Centro Nacional de Biotecnologia , CSIC
+# * [1] SciLifeLab, Stockholm University
 # *
 # * This program is free software; you can redistribute it and/or modify
 # * it under the terms of the GNU General Public License as published by
-# * the Free Software Foundation; either version 2 of the License, or
+# * the Free Software Foundation; either version 3 of the License, or
 # * (at your option) any later version.
 # *
 # * This program is distributed in the hope that it will be useful,
@@ -24,17 +24,8 @@
 # *
 # **************************************************************************
 
-import traceback
+import metadata
+import image
 
-# FIXME: This is a bypass until xmipp is replaced to xmippLib
-# FIXME:   in all imports related to the C++ binding
 
-print("\n >>> WARNING: 'import xmipp' is deprecated for the xmipp binding. "
-      "Please, use 'import xmippLib'")
-stackList = traceback.extract_stack()
-for stackLine in stackList:
-    if 'import xmipp' in stackLine[3] or 'from xmipp import' in stackLine[3]:
-        print("  > Please change 'import xmipp' to 'import xmippLib' in\n"
-              "    File: '%s', line %d.\n" % (stackLine[0], stackLine[1]))
 
-from xmippLib import *
