+++ pwem/emlib/metadata/__init__.py
@@ -1,12 +1,12 @@
 # **************************************************************************
 # *
-# * Authors:     J.M. De la Rosa Trevin (delarosatrevin@scilifelab.se) [1]
+# * Authors:     J.M. de la Rosa Trevin (delarosatrevin@scilifelab.se) [1]
 # *
 # * [1] SciLifeLab, Stockholm University
 # *
 # * This program is free software; you can redistribute it and/or modify
 # * it under the terms of the GNU General Public License as published by
-# * the Free Software Foundation; either version 2 of the License, or
+# * the Free Software Foundation; either version 3 of the License, or
 # * (at your option) any later version.
 # *
 # * This program is distributed in the hope that it will be useful,
@@ -24,13 +24,4 @@
 # *
 # **************************************************************************
 
-from .utils import getSubsetByDefocus, downloadPdb
-from ..emlib.image import ImageHandler, DT_FLOAT
-from .atom_struct import AtomicStructHandler, toCIF, cifToPdb
-from .headers import Ccp4Header
-from .symmetry import getSymmetryMatrices, Icosahedron
-from .transformations import identity_matrix
-from .sequence import *
-from .transformations import (euler_matrix, translation_matrix,
-                              concatenate_matrices)
-
+from .utils import *
