--- pwem/metadata/functions.py
@@ -1,67 +1,0 @@
-# **************************************************************************
-# *
-# * Authors:     J.M. De la Rosa Trevin (jmdelarosa@cnb.csic.es)
-# *
-# * Unidad de  Bioinformatica of Centro Nacional de Biotecnologia , CSIC
-# *
-# * This program is free software; you can redistribute it and/or modify
-# * it under the terms of the GNU General Public License as published by
-# * the Free Software Foundation; either version 2 of the License, or
-# * (at your option) any later version.
-# *
-# * This program is distributed in the hope that it will be useful,
-# * but WITHOUT ANY WARRANTY; without even the implied warranty of
-# * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-# * GNU General Public License for more details.
-# *
-# * You should have received a copy of the GNU General Public License
-# * along with this program; if not, write to the Free Software
-# * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA
-# * 02111-1307  USA
-# *
-# *  All comments concerning this program package may be sent to the
-# *  e-mail address 'scipion@cnb.csic.es'
-# *
-# **************************************************************************
-"""
-Add functions related to metadata
-"""
-import xmippLib
-
-getBlocksInMetaDataFile = xmippLib.getBlocksInMetaDataFile
-
-label2Str = xmippLib.label2Str
-
-colorStr = xmippLib.colorStr
-
-labelType = xmippLib.labelType
-
-labelHasTag = xmippLib.labelHasTag
-
-labelIsImage = xmippLib.labelIsImage
-
-str2Label = xmippLib.str2Label
-
-isValidLabel = xmippLib.isValidLabel
-
-MDValueRelational = xmippLib.MDValueRelational
-
-MDValueEQ = xmippLib.MDValueEQ
-
-MDValueNE = xmippLib.MDValueNE
-
-MDValueLT = xmippLib.MDValueLT
-
-MDValueLE = xmippLib.MDValueLE
-
-MDValueGT = xmippLib.MDValueGT
-
-MDValueGE = xmippLib.MDValueGE
-
-MDValueRange = xmippLib.MDValueRange
-
-addLabelAlias = xmippLib.addLabelAlias
-
-activateMathExtensions = xmippLib.activateMathExtensions
-
-getSymmetryMatrices = xmippLib.SymList().getSymmetryMatrices
